Fix(App and move): Fix shoot
idk if it works but i think

+++ R-TypeGame/Systems/Move/MoveSystemDown/MoveSystemDown.cpp
@@ -69,9 +69,9 @@ void MoveSystemDown::updateDownPosition(ECS::Registry& entityManager, int idxPac
             position->y += speed->speedY;
 
         std::vector<std::any> valuesMoveEntity = {};
-        valuesMoveEntity.push_back(idxPacketEntities);
-        valuesMoveEntity.push_back(position->x);
-        valuesMoveEntity.push_back(position->y);
+        valuesMoveEntity.push_back((int) idxPacketEntities);
+        valuesMoveEntity.push_back((float) position->x);
+        valuesMoveEntity.push_back((float) position->y);
         std::shared_ptr<IEvent> eventMoveEntity = std::make_shared<AEvent>("MoveEntity", valuesMoveEntity);
         entityManager.addEvent(eventMoveEntity);
     } catch(const std::exception& e) {
