+++ R-TypeGame/Systems/Move/MoveSystemUp/MoveSystemUp.cpp
@@ -48,9 +48,9 @@ void MoveSystemUp::updateUpPosition(ECS::Registry& entityManager, int idxPacketE
             position->y -= speed->speedY;
 
         std::vector<std::any> valuesMoveEntity = {};
-        valuesMoveEntity.push_back(idxPacketEntities);
-        valuesMoveEntity.push_back(position->x);
-        valuesMoveEntity.push_back(position->y);
+        valuesMoveEntity.push_back((int) idxPacketEntities);
+        valuesMoveEntity.push_back((float) position->x);
+        valuesMoveEntity.push_back((float) position->y);
         std::shared_ptr<IEvent> eventMoveEntity = std::make_shared<AEvent>("MoveEntity", valuesMoveEntity);
         entityManager.addEvent(eventMoveEntity);
     } catch(const std::exception& e) {
